pyproject.toml
[project]
name = "ai-for-couples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.0",
    "fastapi>=0.124.4",
    "greenlet>=3.3.0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]
